validate db schema to match jpa entities; use BIGINT for timestamp in millis

+++ src/main/resources/db/migration/V1__init.sql
@@ -1,7 +1,7 @@
 CREATE TABLE ponies (
   name VARCHAR(50) NOT NULL,
   available BOOLEAN DEFAULT true,
-  returned_timestamp INT DEFAULT -1,
+  returned_timestamp BIGINT DEFAULT -1,
   recharge_time_mins INT DEFAULT 15,
   PRIMARY KEY (name)
 );
